Move `admin profile` to `support profile` (#3972)
As part of reorganization of support related commands
