+++ cmd/admin-profile.go
@@ -19,6 +19,7 @@ package cmd
 
 import (
 	"github.com/minio/cli"
+	"github.com/minio/pkg/console"
 )
 
 var adminProfileSubcommands = []cli.Command{
@@ -35,10 +36,11 @@ var adminProfileCmd = cli.Command{
 	Flags:           globalFlags,
 	Subcommands:     adminProfileSubcommands,
 	HideHelpCommand: true,
+	Hidden:          true,
 }
 
 // mainAdminProfile is the handle for "mc admin profile" command.
 func mainAdminProfile(ctx *cli.Context) error {
-	commandNotFound(ctx, adminProfileSubcommands)
+	console.Infoln("Please use 'mc support profile'")
 	return nil
 }
